Refactor login and authentication
