+++ api/login.go
@@ -3,7 +3,6 @@ package api
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"github.com/soundrussian/go-practicum-diploma/auth"
 	"github.com/soundrussian/go-practicum-diploma/model"
 	"github.com/soundrussian/go-practicum-diploma/pkg/logging"
@@ -21,8 +20,8 @@ func (api *API) HandleLogin(w http.ResponseWriter, r *http.Request) {
 	var err error
 
 	_, logger := logging.CtxLogger(r.Context())
-	logger = logger.With().Str(logging.HandlerNameKey, "register").Logger()
-	logger.Info().Msg("handling register")
+	logger = logger.With().Str(logging.HandlerNameKey, "login").Logger()
+	logger.Info().Msg("handling login")
 
 	decoder := json.NewDecoder(r.Body)
 	defer r.Body.Close()
@@ -51,26 +50,5 @@ func (api *API) HandleLogin(w http.ResponseWriter, r *http.Request) {
 
 	logger.Info().Msgf("authenticated user id=%d with login %s", user.ID, user.Login)
 
-	var token *string
-	if token, err = api.authService.AuthToken(r.Context(), user); err != nil {
-		logger.Err(err).Msg("failed to get auth token for user")
-		http.Error(w, "user has been registered, but failed to log in", http.StatusInternalServerError)
-		return
-	}
-
-	cookie := &http.Cookie{
-		Name:  "jwt",
-		Value: *token,
-	}
-	http.SetCookie(w, cookie)
-	w.Header().Set("Content-Type", "application/json")
-	w.Header().Set("Authentication", fmt.Sprintf("Bearer %s", *token))
-
-	response := authenticateJSONResponse{Token: *token}
-	encoder := json.NewEncoder(w)
-	if err = encoder.Encode(&response); err != nil {
-		logger.Err(err).Msgf("failed to encode json response from %+v", response)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
+	api.authenticate(user, w, r)
 }
