+++ api/register.go
@@ -3,7 +3,6 @@ package api
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"github.com/soundrussian/go-practicum-diploma/auth"
 	"github.com/soundrussian/go-practicum-diploma/model"
 	"github.com/soundrussian/go-practicum-diploma/pkg/logging"
@@ -13,10 +12,6 @@ import (
 type registerJSONRequest struct {
 	Login    string `json:"login"`
 	Password string `json:"password"`
-}
-
-type authenticateJSONResponse struct {
-	Token string `json:"token"`
 }
 
 func (api *API) HandleRegister(w http.ResponseWriter, r *http.Request) {
@@ -55,26 +50,5 @@ func (api *API) HandleRegister(w http.ResponseWriter, r *http.Request) {
 
 	logger.Info().Msgf("saved user id=%d with login %s", user.ID, user.Login)
 
-	var token *string
-	if token, err = api.authService.AuthToken(r.Context(), user); err != nil {
-		logger.Err(err).Msg("failed to get auth token for user")
-		http.Error(w, "user has been registered, but failed to log in", http.StatusInternalServerError)
-		return
-	}
-
-	cookie := &http.Cookie{
-		Name:  "jwt",
-		Value: *token,
-	}
-	http.SetCookie(w, cookie)
-	w.Header().Set("Content-Type", "application/json")
-	w.Header().Set("Authentication", fmt.Sprintf("Bearer %s", *token))
-
-	response := authenticateJSONResponse{Token: *token}
-	encoder := json.NewEncoder(w)
-	if err = encoder.Encode(&response); err != nil {
-		logger.Err(err).Msgf("failed to encode json response from %+v", response)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
+	api.authenticate(user, w, r)
 }
